deployment/update-db/geocity/requirements.txt
pandas
psycopg2-binary
geopy
locationtagger
spacy
psycopg2-binary